refactor: remove horrid and useless nested directories
