+++ projects/mtd/src/app/pages/about/about-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { AboutComponent } from './about/about.component';
+import { AboutComponent } from './about.component';
 
 const routes: Routes = [
   {
