+++ projects/mtd/src/app/pages/about/about.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles-variables';
+@import '../../../styles-variables';
 
 .floating-img {
   float: right;
