+++ projects/mtd/src/app/pages/about/about.component.ts
@@ -1,13 +1,13 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { ROUTE_ANIMATIONS_ELEMENTS } from '../../../core/core.module';
-import { PronunciationGuideComponent } from '../../../shared/pronunciation-guide/pronunciation-guide.component';
+import { ROUTE_ANIMATIONS_ELEMENTS } from '../../core/core.module';
+import { PronunciationGuideComponent } from '../../shared/pronunciation-guide/pronunciation-guide.component';
 import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'mtd-about',
   templateUrl: './about.component.html',
   styleUrls: [
-    '../../../shared/layout/single/single.component.scss',
+    '../../shared/layout/single/single.component.scss',
     './about.component.scss'
   ],
   changeDetection: ChangeDetectionStrategy.OnPush
