+++ projects/mtd/src/app/pages/bookmarks/bookmarks.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { BookmarksComponent } from './bookmarks/bookmarks.component';
+import { BookmarksComponent } from './bookmarks.component';
 import { SharedModule } from '../../shared/shared.module';
 import { BookmarksRoutingModule } from './bookmarks-routing.module';
