+++ projects/mtd/src/app/pages/browse/browse-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { BrowseComponent } from './browse/browse.component';
+import { BrowseComponent } from './browse.component';
 
 const routes: Routes = [
   {
