+++ projects/mtd/src/app/pages/browse/browse.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles-variables';
+@import '../../../styles-variables';
 
 /* FIXME: Need to calculate this dynamically somehow */
 $mini-alphabet-height: 108px;
