+++ projects/mtd/src/app/pages/browse/browse.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { BrowseComponent } from './browse/browse.component';
+import { BrowseComponent } from './browse.component';
 import { SharedModule } from '../../shared/shared.module';
 import { BrowseRoutingModule } from './browse-routing.module';
