+++ projects/mtd/src/app/pages/home/home.component.ts
@@ -1,11 +1,11 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { ROUTE_ANIMATIONS_ELEMENTS } from '../../../core/core.module';
+import { ROUTE_ANIMATIONS_ELEMENTS } from '../../core/core.module';
 
 @Component({
   selector: 'mtd-home',
   templateUrl: './home.component.html',
   styleUrls: [
-    '../../../shared/layout/single/single.component.scss',
+    '../../shared/layout/single/single.component.scss',
     './home.component.scss'
   ],
   changeDetection: ChangeDetectionStrategy.OnPush
