+++ projects/mtd/src/app/pages/people/people.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles-variables';
+@import '../../../styles-variables';
 
 @mixin images($width) {
   &__left {
