+++ projects/mtd/src/app/pages/people/people.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { SharedModule } from '../../shared/shared.module';
 
-import { PeopleComponent } from './people/people.component';
+import { PeopleComponent } from './people.component';
 import { PeopleRoutingModule } from './people-routing.module';
 
 @NgModule({
