+++ projects/mtd/src/app/pages/random/random.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles-variables';
+@import '../../../styles-variables';
 
 /* Match the height of the h1 we would normally have in mtd-single */
 @media (max-width: map-get($angular-breakpoints, tablet) - 1) {
