+++ projects/mtd/src/app/pages/random/random.component.scss-theme.scss
