+++ projects/mtd/src/app/pages/random/random.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RandomComponent } from './random/random.component';
+import { RandomComponent } from './random.component';
 import { SharedModule } from '../../shared/shared.module';
 import { RandomRoutingModule } from './random-routing.module';
