+++ projects/mtd/src/app/pages/search/search-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { SearchComponent } from './search/search.component';
+import { SearchComponent } from './search.component';
 
 const routes: Routes = [
   {
