+++ projects/mtd/src/app/pages/search/search.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles-variables';
+@import '../../../styles-variables';
 
 .page__content {
   max-height: calc($notphone-content-height - $notphone-margin);
