+++ projects/mtd/src/app/pages/search/search.component.scss-theme.scss
