+++ projects/mtd/src/app/pages/search/search.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SearchComponent } from './search/search.component';
+import { SearchComponent } from './search.component';
 import { SharedModule } from '../../shared/shared.module';
 import { SearchRoutingModule } from './search-routing.module';
