+++ projects/mtd/src/app/pages/settings/settings-container.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles-variables';
+@import '../../../styles-variables';
 
 h1 {
   text-align: center;
