+++ projects/mtd/src/app/pages/settings/settings-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { SettingsContainerComponent } from './settings/settings-container.component';
+import { SettingsContainerComponent } from './settings-container.component';
 
 const routes: Routes = [
   {
