+++ projects/mtd/src/app/pages/settings/settings.module.ts
@@ -4,7 +4,7 @@ import { CommonModule } from '@angular/common';
 import { SharedModule } from '../../shared/shared.module';
 
 import { SettingsRoutingModule } from './settings-routing.module';
-import { SettingsContainerComponent } from './settings/settings-container.component';
+import { SettingsContainerComponent } from './settings-container.component';
 
 @NgModule({
   declarations: [SettingsContainerComponent],
