+++ projects/mtd/src/app/pages/speakers/speakers.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles-variables';
+@import '../../../styles-variables';
 
 .floating-img {
   float: right;
