+++ projects/mtd/src/styles.scss
@@ -21,10 +21,10 @@
 /* These are all mostly the same thing, unfortunately */
 @import './app/app/app.component.scss-theme';
 @import './app/shared/big-input/big-input/big-input.component.scss-theme';
-@import './app/pages/bookmarks/bookmarks/bookmarks.component.scss-theme.scss';
-@import './app/pages/browse/browse/browse.component.scss-theme.scss';
-@import './app/pages/random/random/random.component.scss-theme.scss';
-@import './app/pages/search/search/search.component.scss-theme.scss';
+@import './app/pages/bookmarks/bookmarks.component.scss-theme.scss';
+@import './app/pages/browse/browse.component.scss-theme.scss';
+@import './app/pages/random/random.component.scss-theme.scss';
+@import './app/pages/search/search.component.scss-theme.scss';
 @import './app/shared/layout/single/single.component.scss-theme.scss';
 @import './app/shared/entry-list/entry-list.component.scss-theme.scss';
 @import './app/shared/word-modal/word-modal.component.scss-theme.scss';
